+++ pie-charts/src/main/java/com/ivy/pie_charts/PieChartStatisticScreen.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.ivy.base.CategoryAmount
+import com.ivy.pie_charts.model.CategoryAmount
 import com.ivy.base.IvyWalletPreview
 import com.ivy.base.TimePeriod
 import com.ivy.base.ivyWalletCtx
